+++ core/lib/snt/core/services/error.rb
@@ -1,33 +1,17 @@
 module SNT
   module Core
     module Services
-      class Error
+      class Error < Message
         VALIDATION_ERROR = 'VALIDATION_ERROR'.freeze
 
-        attr_accessor :code, :message
-
-        def initialize(args)
-          args.each { |key, value| instance_variable_set("@#{key}", value) }
-        end
-
         def self.new_error(code, obj)
-          message =
-            if obj.is_a?(Symbol)
-              I18n.t(obj)
-            elsif obj.is_a?(String)
-              obj
-            end
-
-          new(code: code, message: message)
+          new_message(code, obj)
         end
 
         def self.new_validation_error(obj)
           new_error(VALIDATION_ERROR, obj)
         end
 
-        def to_hash
-          { code: code, message: message }
-        end
       end
     end
   end
