+++ core/lib/snt/core/services/message.rb
@@ -1,0 +1,29 @@
+module SNT
+  module Core
+    module Services
+      class Message
+        
+        attr_accessor :code, :message
+
+        def initialize(args)
+          args.each { |key, value| instance_variable_set("@#{key}", value) }
+        end
+
+        def self.new_message(code, obj)
+          message =
+            if obj.is_a?(Symbol)
+              I18n.t(obj)
+            elsif obj.is_a?(String)
+              obj
+            end
+
+          new(code: code, message: message)
+        end
+
+        def to_hash
+          { code: code, message: message }
+        end
+      end
+    end
+  end
+end
