+++ core/lib/snt/core/services/warning.rb
@@ -1,0 +1,13 @@
+module SNT
+  module Core
+    module Services
+      class Warning < Message
+
+        def self.new_warning(code, obj)
+          new_message(code, obj)
+        end
+
+      end
+    end
+  end
+end
